fix: refactored admin status tracking
Refactored the way the project tracks the admin status of the user by creating a UserStatus class which contains the isAdmin variable which can be set to true and false appropriately. Using this class I implemented a feature for the buttons to dynamically appear/disappear if the user is an admin or not.